+++ packages/syntax-core/src/Button/Button.tsx
@@ -83,7 +83,7 @@ type ButtonType = {
   /**
    * The callback to be called when the button is clicked
    */
-  onClick: React.MouseEventHandler<HTMLButtonElement>;
+  onClick?: React.MouseEventHandler<HTMLButtonElement>;
   /**
    * The tooltip to be displayed when the user hovers over the button
    */
@@ -94,6 +94,9 @@ type ButtonType = {
   type?: "button" | "submit" | "reset";
 };
 
+/**
+ * [Button](https://cambly-syntax.vercel.app/?path=/docs/components-button--docs) is used to trigger an action or event, such as submitting a form, opening a dialog, canceling an action, or performing a delete operation.
+ */
 const Button = forwardRef<HTMLButtonElement, ButtonType>(
   (
     {